test(#33): controllerTest 통합
-Cors Filter 때문에 WebMvcTest로 진행이 되지 않음.
